Monkeypatch bigquery and gcs at module load time to avoid always loading bigquery/gcs and all its dependencies.
